+++ src/LanguageServer.spec.ts
@@ -877,6 +877,14 @@ describe('LanguageServer', () => {
             }
         });
     });
+
+    describe('CustomCommands', () => {
+        describe('TranspileFile', () => {
+            it('returns pathAbsolute to support backwards compatibility', () => {
+
+            });
+        });
+    });
 });
 
 export function getFileProtocolPath(fullPath: string) {