+++ src/LanguageServer.ts
@@ -1102,7 +1102,9 @@ export class LanguageServer {
     public async onExecuteCommand(params: ExecuteCommandParams) {
         await this.waitAllProgramFirstRuns();
         if (params.command === CustomCommands.TranspileFile) {
-            return this.transpileFile(params.arguments[0]);
+            const result = await this.transpileFile(params.arguments[0]);
+            //back-compat: include `pathAbsolute` property so older vscode versions still work
+            (result as any).pathAbsolute = result.srcPath;
         }
     }