+++ src/Program.ts
@@ -1275,6 +1275,7 @@ export class Program {
         const editor = new AstEditor();
 
         this.plugins.emit('beforeFileTranspile', {
+            program: this,
             file: file,
             outputPath: outputPath,
             editor: editor
@@ -1296,6 +1297,7 @@ export class Program {
         }
 
         const event: AfterFileTranspileEvent = {
+            program: this,
             file: file,
             outputPath: outputPath,
             editor: editor,
@@ -1309,7 +1311,6 @@ export class Program {
         editor.undoAll();
 
         return {
-            pathAbsolute: file.srcPath,
             srcPath: file.srcPath,
             pkgPath: file.pkgPath,
             code: event.code,
@@ -1325,7 +1326,7 @@ export class Program {
             return collection;
         }, {});
 
-        const entries = Object.values(this.files).map(file => {
+        const getOutputPath = (file: BscFile) => {
             let filePathObj = mappedFileEntries[s`${file.srcPath}`];
             if (!filePathObj) {
                 //this file has been added in-memory, from a plugin, for example
@@ -1339,23 +1340,19 @@ export class Program {
             let outputPath = filePathObj.dest.replace(/\.bs$/gi, '.brs');
             //prepend the staging folder path
             outputPath = s`${stagingFolderPath}/${outputPath}`;
-            return {
-                file: file,
-                outputPath: outputPath
-            };
-        });
+            return outputPath;
+        };
 
-        const astEditor = new AstEditor();
+        const processedFiles = new Set<File>();
 
-        this.plugins.emit('beforeProgramTranspile', this, entries, astEditor);
+        const transpileFile = async (file: BscFile, outputPath?: string) => {
+            //mark this file as processed so we don't do it again
+            processedFiles.add(file);
 
-        const promises = entries.map(async (entry) => {
             //skip transpiling typedef files
-            if (isBrsFile(entry.file) && entry.file.isTypedef) {
+            if (isBrsFile(file) && file.isTypedef) {
                 return;
             }
-
-            const { file, outputPath } = entry;
 
             const fileTranspileResult = this._getTranspiledFileContents(file, outputPath);
 
@@ -1375,13 +1372,47 @@ export class Program {
                 const typedefPath = outputPath.replace(/\.brs$/i, '.d.bs');
                 await fsExtra.writeFile(typedefPath, fileTranspileResult.typedef);
             }
+        };
+
+        const entries = Object.values(this.files).map(file => {
+            return {
+                file: file,
+                outputPath: getOutputPath(file)
+            };
+        });
+
+        const astEditor = new AstEditor();
+
+        this.plugins.emit('beforeProgramTranspile', this, entries, astEditor);
+
+        let promises = entries.map(async (entry) => {
+            return transpileFile(entry.file, entry.outputPath);
         });
 
         //if there's no bslib file already loaded into the program, copy it to the staging directory
-        if (!this.getFileByPkgPath(bslibAliasedRokuModulesPkgPath) && !this.getFileByPkgPath(s`source/bslib.brs`)) {
+        if (!this.getFile(bslibAliasedRokuModulesPkgPath) && !this.getFile(s`source/bslib.brs`)) {
             promises.push(util.copyBslibToStaging(stagingFolderPath));
         }
         await Promise.all(promises);
+
+        //transpile any new files that plugins added since the start of this transpile process
+        do {
+            promises = [];
+            for (const key in this.files) {
+                const file = this.files[key];
+                //this is a new file
+                if (!processedFiles.has(file)) {
+                    promises.push(
+                        transpileFile(file, getOutputPath(file))
+                    );
+                }
+            }
+            if (promises.length > 0) {
+                this.logger.info(`Transpiling ${promises.length} new files`);
+                await Promise.all(promises);
+            }
+        }
+        while (promises.length > 0);
 
         this.plugins.emit('afterProgramTranspile', this, entries, astEditor);
         astEditor.undoAll();
@@ -1461,10 +1492,6 @@ export class Program {
 
 export interface FileTranspileResult {
     srcPath: string;
-    /**
-     * @deprecated use `srcPath`
-     */
-    pathAbsolute: string;
     pkgPath: string;
     code: string;
     map: SourceMapGenerator;